chore: tidy up files
Add schemas.ts
Add errors.ts
Add final TODOs
Note copied code
