+++ plugins/maciVoting/services/coordinator/config.ts
@@ -1,8 +1,11 @@
 import { EPolicies, EMode, EInitialVoiceCreditProxies } from "@maci-protocol/sdk";
-import { startDate, pollDuration, coordinatorMACIKeypair } from "./coordinator";
+import { coordinatorMACIKeypair } from "./coordinator";
 import { type IDeployPollConfig } from "./utils/types";
 
 const MSG_BATCH_SIZE = 20;
+
+const startDate = Math.floor(Date.now() / 1000) + 100;
+const pollDuration = 60; // TODO:
 
 export const pollDeploymentConfig: IDeployPollConfig = {
   startDate,
@@ -14,10 +17,10 @@ export const pollDeploymentConfig: IDeployPollConfig = {
   pollStateTreeDepth: 10,
   voteOptionTreeDepth: 2,
   policy: {
-    type: EPolicies.FreeForAll,
+    type: EPolicies.FreeForAll, // TODO:
   },
   initialVoiceCreditsProxy: {
-    type: EInitialVoiceCreditProxies.Constant,
+    type: EInitialVoiceCreditProxies.Constant, // TODO:
     args: {
       amount: 100,
     },
